% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_min_tier.R
\name{add_MinimalTier}
\alias{add_MinimalTier}
\title{Converts a basic GAT2 Transcript to an Mininmal transcript}
\usage{
add_MinimalTier(file)
}
\arguments{
\item{file}{path of an exb file}
}
\value{
NULL saves file to original path with suffix _M;in
}
\description{
Converts a basic GAT2 Transcript to an Mininmal transcript
}
