% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_min_tier.R
\name{AddMinTier}
\alias{AddMinTier}
\title{Converts a basic GAT2 Transcript to an Mininmal transcript}
\usage{
AddMinTier(file)
}
\arguments{
\item{file}{}
}
\value{

}
\description{
Converts a basic GAT2 Transcript to an Mininmal transcript
}
