% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Exb.R
\name{check_exb}
\alias{check_exb}
\title{Check Filetype}
\usage{
check_exb(path)
}
\arguments{
\item{path}{Needs the path of the file you want to read}
}
\value{
Returns logical value
}
\description{
Check Filetype
}
\examples{
check_exb(path)
}
