% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_transcription_status.R
\name{corpus_transcription_status}
\alias{corpus_transcription_status}
\title{corpus_transcription_status}
\usage{
corpus_transcription_status(path_files, subDirs = FALSE, summarize = TRUE)
}
\arguments{
\item{summarize}{}
}
\value{

}
\description{
corpus_transcription_status
}
\examples{

}
