% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_exb_files.R
\name{merge_exb_files}
\alias{merge_exb_files}
\title{merge_exb_files
Merges several exb files into one xml elemen}
\usage{
merge_exb_files(files = list())
}
\arguments{
\item{files}{}
}
\value{
returns a xml element
}
\description{
merge_exb_files
Merges several exb files into one xml elemen
}
