% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_events_exb.R
\name{read_events}
\alias{read_events}
\title{read_events()}
\usage{
read_events(file, name)
}
\arguments{
\item{file}{xml Object from an .exb transcription file}

\item{path}{path of the .exb transcription file}
}
\value{
Returns data.frame containing all Information; events are not sortetd
}
\description{
Reads events from an .exb transcription file
}
\examples{
 path <- system.file("extdata", "Example_linear.exb", package = "ExmaraldaR", mustWork = TRUE)
 file <- xml2::read_xml(path, encoding="UTF-8")
read_events(file, path)
}
