% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exb_dir.R
\name{read_exb_dir}
\alias{read_exb_dir}
\title{Title}
\usage{
read_exb_dir(
  pathDir,
  readAnn = TRUE,
  addDescription = FALSE,
  addMetaData = FALSE,
  addIPNumber = TRUE,
  IPEndSign = c("|", ".", ";", ",", ",", "?", "=", "-")
)
}
\arguments{
\item{pathDir}{path of directory with .exb files}

\item{readAnn}{logical, if annotation tiers should be read}

\item{addDescription}{logical, if description tiers should be read NOT YET POSSIBLE}

\item{addMetaData}{logical, if information from speaker table should be read}

\item{addIPNumber}{logical, if there should be an nummeration of intonation phrases}

\item{IPEndSign}{character, spezifies characters that indicate the end of a intonation unit}
}
\value{
data.frame
}
\description{
Title
}
