% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exb_sciebo.R
\name{read_exb_xml}
\alias{read_exb_xml}
\title{Title}
\usage{
read_exb_xml(
  file,
  filename,
  readAnn = TRUE,
  annotation = c("linear", "multilayer"),
  addDescription = FALSE,
  addMetaData = FALSE,
  sortMetaData = TRUE
)
}
\arguments{
\item{addDescription}{logical value wheter description tiers should be inclouded}
}
\value{

}
\description{
Title
}
