% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Title}
\usage{
read_metadata(path)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
Title
}
\examples{
path <- system.file("extdata", "Example_linear.exb", package = "ExmaraldaR", mustWork = TRUE)
read_metadata(path)
}
