% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_timeline.R
\name{read_timeline}
\alias{read_timeline}
\title{read_timeline()}
\usage{
read_timeline(file)
}
\arguments{
\item{file}{Requires the xml-object of an .exb-file}
}
\value{
Returns a Vector of the timeline in an EXMARaLDA transcription file
}
\description{
Reads timestamps from an exb Transcription file
}
\examples{
read_timeline(file)
}
