% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_annotations_multilayer.R
\name{sort_annotations_multilayer}
\alias{sort_annotations_multilayer}
\title{Title}
\usage{
sort_annotations_multilayer(file, AnnotationTiers, events_sorted)
}
\arguments{
\item{events_sorted}{}
}
\value{

}
\description{
Title
}
\examples{
path <- system.file("extdata", "Example_multi.exb", package = "ExmaraldaR", mustWork = TRUE)
events_sorted <- sort_events(events= read_events(file, path), timeline= read_timeline(file)) \%>\% add_IpNumber(events_sorted)
sort_annoations_multilayer(file= xml2::read_xml(path, AnnotationTiers= xml2::xml_find_all(file,".//tier[@type='a']"), encoding="UTF-8"), events_sorted)
}
