% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_annotations_linear.R
\name{sort_anntotations_linear}
\alias{sort_anntotations_linear}
\title{Title}
\usage{
sort_anntotations_linear(exb)
}
\arguments{
\item{exb}{}
}
\value{

}
\description{
Title
}
\examples{
path <- system.file("extdata", "Example_linear.exb", package = "ExmaraldaR", mustWork = TRUE) # for a linear annotation
     file <- xml2::read_xml(path, encoding="UTF-8")
     timeline <- read_timeline(file)
     events <- read_events(file, path)
 events_sorted <- sort_events(events, timeline)
 events_sorted <- add_IpNumber(events_sorted)
 annotations <- read_annotations_linear(file= xml2::read_xml(path, encoding="UTF-8"))
 exb <- dplyr::left_join(events_sorted, annotations,by = c("Speaker", "Start", "End"))
 MultiEventAnn <- dplyr::anti_join( annotations,events_sorted, by=c("Speaker", "Start", "End")) # check for annotations for more than 1 event
 if(nrow(MultiEventAnn)!=0){
 for (n in 1:nrow(MultiEventAnn)) {
 a <- which(events_sorted[,'Start']==MultiEventAnn[n, 'Start'])
 b <- which(events_sorted[,'End']==MultiEventAnn[n, 'End'])
 exb[a:b,colnames(MultiEventAnn)[ncol(MultiEventAnn)]] <- MultiEventAnn[n, ncol(MultiEventAnn)]
 }
 }
 sort_anntotations_linear(exb)
}
