% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_by_unit.R
\name{sort_by_unit}
\alias{sort_by_unit}
\title{#' sort_by_unit
#'
#' @param exb needs a dataframe as given by read_exb_file or read_exb_dir
#' @param Unit_column column where the unit is defined. Needs to be consisted. Default is 1 as the output of read_exb_file and read_exb_dir saves the IP number there
#' @param noObservationAsNA logical value, whether empty cells should be counted as NA or zero
#' @param drop logical value, whether rows without an observation should be dropped or not
#' @param percentage logical value whether values per Variable per Unite should be relative or absolute
#'
#'
#' @return returns data.frame
#' @export
#'
#' @examples
#' path <- system.file("extdata", "Example_linear.exb", package = "ExmaraldaR", mustWork = TRUE) # for a linear annotation
#' example_linear <- read_exb_file(path, readAnn = TRUE, annotation= "linear", addMetaData = TRUE)
#' sort_by_unit(example_linear, Unit_column=1, drop= TRUE, percentage= TRUE)
#'}
\usage{
sort_by_unit(
  exb,
  Unit_column = 1,
  noObservationAsNA = TRUE,
  drop = FALSE,
  percentage = TRUE
)
}
\description{
#' sort_by_unit
#'
#' @param exb needs a dataframe as given by read_exb_file or read_exb_dir
#' @param Unit_column column where the unit is defined. Needs to be consisted. Default is 1 as the output of read_exb_file and read_exb_dir saves the IP number there
#' @param noObservationAsNA logical value, whether empty cells should be counted as NA or zero
#' @param drop logical value, whether rows without an observation should be dropped or not
#' @param percentage logical value whether values per Variable per Unite should be relative or absolute
#'
#'
#' @return returns data.frame
#' @export
#'
#' @examples
#' path <- system.file("extdata", "Example_linear.exb", package = "ExmaraldaR", mustWork = TRUE) # for a linear annotation
#' example_linear <- read_exb_file(path, readAnn = TRUE, annotation= "linear", addMetaData = TRUE)
#' sort_by_unit(example_linear, Unit_column=1, drop= TRUE, percentage= TRUE)
#'
}
