% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewEvent.R
\name{view_event}
\alias{view_event}
\title{view_event
View event in Praat}
\usage{
view_event(exb, rowEvent = 1, pathFileDir = character(0), pathPraat = getwd())
}
\arguments{
\item{exb}{exb object as created by exb_read_file or exb_read_dir; can also be a subset}

\item{rowEvent}{row of the event you want to play}

\item{pathFileDir}{path of the audio file dir (, to provide dir is useful if there are several audio files in one corpus;has to be .wav at the moment)}

\item{pathPraat}{path of Praat.exe; if not provided it will be assumend praat is in the working directory}
}
\description{
view_event
View event in Praat
}
