% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_back_dir.R
\name{write_back_dir}
\alias{write_back_dir}
\title{write_back_dir
Writes back edited annotations to original files
needs csv file form data.frame created by red_exb_dir()}
\usage{
write_back_dir(
  exb,
  sep = ",",
  PathExbDir,
  PathNewFiles = PathExbDir,
  suffix = "_new",
  verbose = TRUE,
  transcription_text = "Text",
  annotation_colums = NA,
  overwrite_annotations = TRUE,
  assignSpeakersAnnotation = FALSE,
  recreate_timeline = FALSE,
  hideAnnotationTiers = FALSE
)
}
\arguments{
\item{exb}{data.frame created by read_exb_dir() or csv file created by such an object}

\item{sep}{Seperator in the csv file, if object can be left out}

\item{PathExbDir}{Path of the original files. Attention: the files must have the same names as in the "File" collumn}

\item{PathNewFiles}{Default is the original Directory, can be changend}

\item{suffix}{suffix to be added to the new files, default is "_new"}

\item{verbose}{logical, if TRUE, status will be shown}

\item{transcription_text}{specify column with transcription text}

\item{annotation_colums}{specify names of annotaion columns}

\item{overwrite_annotations}{if true old annotaitons will be deleted, AT THE MOMENT ONLY WORKS IF TRUE}

\item{assignSpeakerAnnotation}{if TRUE one annotation tier per category per speaker will be added, AT THME MOMENT ONLY WORKS IF FALSE}
}
\description{
write_back_dir
Writes back edited annotations to original files
needs csv file form data.frame created by red_exb_dir()
}
