% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_IpNumber.R
\name{add_IpNumber}
\alias{add_IpNumber}
\title{add_IpNumber()}
\usage{
add_IpNumber(events_sorted)
}
\arguments{
\item{events_sorted}{Requieres data.frame of sorted events as returned by sort_events()}
}
\value{
Returns a data.frame with sorted events with a conitnuus IP-Number
}
\description{
add_IpNumber()
}
\examples{
add_IpNumber(events_sorted)

}
