% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_annotations_exb.R
\name{read_annotations_linar}
\alias{read_annotations_linar}
\title{read_annotations_linear()}
\usage{
read_annotations_linar(file)
}
\arguments{
\item{file}{Needs the xml object of an exb transcription file}
}
\value{
Returns a data.frame
}
\description{
Reads anntotations from an exb transcription file
}
\examples{
read_annotations(file)
}
