% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_events.R
\name{read_events}
\alias{read_events}
\title{Helper function reads events}
\usage{
read_events(file, path)
}
\arguments{
\item{file}{xml object}

\item{path}{path of file}
}
\description{
Helper function reads events
}
