% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_events_exb.R
\name{read_events}
\alias{read_events}
\title{read_events()}
\usage{
read_events(file, path)
}
\arguments{
\item{file}{xml Object from an .exb transcription file}

\item{path}{path of the .exb transcription file}
}
\value{
Returns data.frame containing all Information; events are not sortetd
}
\description{
Reads events from an .exb transcription file
}
\examples{
read_events(file)
}
