% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exb_dir.R
\name{read_exb_dir}
\alias{read_exb_dir}
\title{Read entire directory of exb files annotated with the same annotation specification}
\usage{
read_exb_dir(
  pathDir,
  addMetaData = FALSE,
  readAnn = TRUE,
  annotation = c("linear", "multilayer"),
  addDescription = FALSE
)
}
\arguments{
\item{pathDir}{path of the directory}

\item{addMetaData}{Logical Value, wheter Metadata should be read from the speakertable}

\item{readAnn}{Logical Value, whetaer annotation tiers should be read and sorted}

\item{annotation}{"linear" or multilayer. See vignette for further information}

\item{addDescription}{logical value wheter description tiers should be inclouded}
}
\value{
returns dataframe that contains all events, an IP number and formatted annotations
}
\description{
Read entire directory of exb files annotated with the same annotation specification
}
\examples{
path <- system.file("extdata\\\\read_dir", package = "ExmaraldaR", mustWork = TRUE)
read_exb_dir(path, addMetaData = TRUE, readAnn = TRUE, annotation = "linear")
}
