% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exb_file.R
\name{read_exb_file}
\alias{read_exb_file}
\title{read_exb_file()}
\usage{
read_exb_file(
  path,
  readAnn = TRUE,
  annotation = c("linear", "multilayer"),
  addMetaData = FALSE,
  sortMetaData = TRUE
)
}
\arguments{
\item{path}{Path of an exb transcription file}

\item{readAnn}{Logical Value, whetaer annotation tiers should be read and sorted}

\item{annotation}{"linear" or multilayer. See vignette for further information}

\item{addMetaData}{Logical Value, wheter Metadata should be read from the speakertable}

\item{sortMetaData}{Logical value, wheter metadata should be sorted directly after the speaker name or at the end}
}
\value{
Returns a data frame that contains the transcription and the annotations
}
\description{
Function that reads an exb transcription file
}
\examples{
path <- system.file("extdata", "Example_linear.exb", package = "ExmaraldaR", mustWork = TRUE) # for a linear annotation
example_linear <- read_exb_file(path, readAnn = TRUE, annotation= "linear", addMetaData = TRUE)
path <- system.file("extdata", "Example_multi.exb", package = "ExmaraldaR", mustWork = TRUE) # for a multilayer annotation
example_multi <- read_exb_file(path, readAnn = TRUE, annotation= "multilayer", addMetaData = TRUE)

}
