% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exb_sciebo.R
\name{read_exb_sciebo}
\alias{read_exb_sciebo}
\title{Title}
\usage{
read_exb_sciebo(
  path_list,
  username,
  password,
  readAnn = TRUE,
  addDescription = FALSE,
  annotation = c("linear", "multilayer"),
  addMetaData = FALSE,
  sortMetaData = TRUE
)
}
\arguments{
\item{addDescription}{logical value wheter description tiers should be inclouded}
}
\description{
Title
}
