% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Reads Speaker information from exb file}
\usage{
read_metadata(path)
}
\arguments{
\item{path}{}
}
\value{
data.frame
}
\description{
Reads Speaker information from exb file
}
