% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{read_metadata
Reads metadata added via the speakertable}
\usage{
read_metadata(file)
}
\arguments{
\item{file}{xml object of a exb file}
}
\value{
Returns a data.table
}
\description{
read_metadata
Reads metadata added via the speakertable
}
\examples{
path <- system.file("extdata", "Example_linear.exb", package = "ExmaraldaR", mustWork = TRUE)
read_metadata(path)
}
