% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_annotations_multilayer.R
\name{sort_annotations_multilayer}
\alias{sort_annotations_multilayer}
\title{Reads and sorts annotations in multilayer format}
\usage{
sort_annotations_multilayer(file, AnnotationTiers, events_sorted)
}
\arguments{
\item{file}{object returned by xml2::read_xml() of an .exb-file}

\item{AnnotationTiers}{vector containing the annotation tiers}

\item{events_sorted}{data.frame returned by sort_events()}
}
\value{

}
\description{
Reads and sorts annotations in multilayer format
}
\examples{
path <- system.file("extdata", "Example_multi.exb", package = "ExmaraldaR", mustWork = TRUE)
file <- xml2::read_xml(path, encoding="UTF-8")
events_sorted <- sort_events(events= read_events(file, path), timeline= read_timeline(file))
events_sorted <- add_IpNumber(events_sorted)
sort_annotations_multilayer(file, AnnotationTiers= xml2::xml_find_all(file,".//tier[@type='a']"), events_sorted)
}
