% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_events.R
\name{sort_events}
\alias{sort_events}
\title{sort_events()}
\usage{
sort_events(events, timeline)
}
\arguments{
\item{events}{List of events as returned by read_events()}

\item{timeline}{List of timestamps returned by read_timeline()}
}
\value{
Returns list of sorted events
}
\description{
sort_events()
}
\examples{
sort_events(events, timeline)
}
