% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_back_dir.R
\name{write_back_dir}
\alias{write_back_dir}
\title{write_back_dir
Writes back edited annotations to original files
needs csv file form data.frame created by red_exb_dir()}
\usage{
write_back_dir(
  exb,
  sep = ",",
  PathExbDir,
  PathNewFiles = PathExbDir,
  suffix = "_new"
)
}
\arguments{
\item{exb}{data.frame created by red_exb_dir() or csv file created by such an object}

\item{sep}{Seperator in the csv file, if object can be left out}

\item{PathExbDir}{Path of the original files. Attention: the files must have the same names as in the "File" collumn}

\item{PathNewFiles}{Default is the original Directory, can be changend}

\item{suffix}{suffix to be added to the new files, default is "_new"}
}
\value{

}
\description{
write_back_dir
Writes back edited annotations to original files
needs csv file form data.frame created by red_exb_dir()
}
